requirements.txt
openai-whisper
pytest
pytest-cov
